Add integration tests against endpoint
