+++ features/bootstrap/Contexts/Domain/GetConfigContext.php
@@ -23,7 +23,7 @@ class GetConfigContext implements Context
 
     public function __construct(string $cassettePath)
     {
-        $config = new \Services\TestConfig($cassettePath);
+        $config = new \Services\HttpMockConfig($cassettePath);
         $this->httpRecorder = $config->httpRecorder();
         $this->application = new Application($config);
     }
