+++ features/bootstrap/Services/TestConfig.php
@@ -6,17 +6,6 @@ use Inviqa\Clearpay\Config;
 
 class TestConfig implements Config
 {
-    /**
-     * @var HttpRecorder
-     */
-    private $httpRecorder;
-
-    public function __construct($cassettePath)
-    {
-        $this->httpRecorder = HttpRecorder::fromShelf($cassettePath);
-        $this->httpRecorder->powerOn();
-    }
-
     public function uri(): string
     {
         return getenv('CLEARPAY_URI');
@@ -30,10 +19,5 @@ class TestConfig implements Config
     public function password(): string
     {
         return getenv('CLEARPAY_PASSWORD');
-    }
-
-    public function httpRecorder(): HttpRecorder
-    {
-        return $this->httpRecorder;
     }
 }
