+++ features/files/create_checkout.feature
@@ -1,3 +1,4 @@
+@http
 Feature: Initiate Clearpay payment process by creating a checkout
 
     Scenario: Create a successful checkout
